+++ src/main/java/lotto/controller/LottoFactoryImpl.java
@@ -1,14 +1,11 @@
 package lotto.controller;
 
 import java.util.List;
-import lotto.model.entity.Lotto;
-import lotto.model.entity.Lotto.LottoBonusNumber;
-import lotto.model.entity.Money;
+import lotto.model.entity.*;
 import lotto.model.vendor.IVendor;
 import lotto.model.vendor.prize.LottoPrizeVendor;
 import lotto.util.ExceptionHandler;
-import lotto.view.input.IInputView;
-import lotto.view.input.InputViewProxy;
+import lotto.view.input.*;
 
 class LottoFactoryImpl implements ILottoFactory {
