+++ src/main/java/lotto/model/entity/Lotto.java
@@ -1,8 +1,6 @@
 package lotto.model.entity;
 
-import java.util.List;
-import java.util.SortedSet;
-import java.util.TreeSet;
+import java.util.*;
 import lotto.util.LottoEnum;
 
 public record Lotto(SortedSet<Integer> numbers) implements Priced {
@@ -56,25 +54,4 @@ public record Lotto(SortedSet<Integer> numbers) implements Priced {
         return contains(bonusNumber.number);
     }
 
-    public static final class LottoBonusNumber {
-        final int number;
-
-        public LottoBonusNumber(Lotto winningLotto, int number) {
-            this.number = number;
-            validate(winningLotto);
-        }
-
-        private void validate(Lotto winningLotto) {
-            int MIN_NUMBER = LottoEnum.MIN_NUMBER.getValue();
-            int MAX_NUMBER = LottoEnum.MAX_NUMBER.getValue();
-            if (number < MIN_NUMBER || number > MAX_NUMBER) {
-                throw new IllegalArgumentException("보너스 번호는 " + MIN_NUMBER + "이상 " + MAX_NUMBER + "이하의 숫자여야 합니다.");
-            }
-
-            if (winningLotto.numbers().contains(number)) {
-                throw new IllegalArgumentException("보너스 번호는 당첨 번호와 중복될 수 없습니다.");
-            }
-        }
-
-    }
 }
