+++ src/main/java/lotto/util/GlobalExceptionMessageEnum.java
@@ -1,8 +1,5 @@
 package lotto.util;
 
-import lombok.Getter;
-
-@Getter
 public enum GlobalExceptionMessageEnum {
     ERROR_MESSAGE("[ERROR]");
     private final String value;
@@ -11,4 +8,7 @@ public enum GlobalExceptionMessageEnum {
         this.value = value;
     }
 
+    public String getValue() {
+        return value;
+    }
 }
