+++ src/main/java/lotto/view/input/InputViewProxy.java
@@ -1,8 +1,6 @@
 package lotto.view.input;
 
-import lotto.model.entity.Lotto;
-import lotto.model.entity.Lotto.LottoBonusNumber;
-import lotto.model.entity.Money;
+import lotto.model.entity.*;
 
 public class InputViewProxy implements IInputView {
