+++ src/test/java/lotto/model/entity/LottoGenerateTest.java
@@ -2,8 +2,7 @@ package lotto.model.entity;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
 
 class LottoGenerateTest {
