+++ src/test/java/lotto/model/entity/MoneyGenerateTest.java
@@ -1,8 +1,7 @@
 package lotto.model.entity;
 
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
 
 class MoneyGenerateTest {
     @DisplayName("돈의 액수가 음수이면 예외가 발생한다.")
